lab5 basic 1

+++ lab_5/shell/include/stdint.h
@@ -1,8 +1,5 @@
-
-typedef unsigned int uint16_t;
-
-typedef unsigned long uint32_t;
-
+typedef unsigned short uint16_t;
+typedef unsigned int uint32_t;
 typedef unsigned long long uint64_t;
 typedef unsigned long uintptr_t;
 typedef float float32;
